+++ haxe/std/objc/ios/map/MKOverlayView.h
@@ -65,4 +65,4 @@ extern class MKOverlayView extends UIView {
 // Road widths are typically not drawn to scale on the map.  This function
 // returns the approximate width in points of roads at the specified zoomScale.
 // The result of this function is suitable for use with CGContextSetLineWidth.
-MK_EXTERN CGFloat MKRoadWidthAtZoomScale(MKZoomScale zoomScale) NS_AVAILABLE(NA, 4_0);
+MK_EXTERN Float MKRoadWidthAtZoomScale(MKZoomScale zoomScale) NS_AVAILABLE(NA, 4_0);
